Refactored TestContainer and Xtress Agent - now works fully on backend. next is to add requests per second followed by testing multiple endpoints within a single test run
