+++ lib/proxyServer.js
@@ -4,7 +4,7 @@ const http = require('http').Server(app);
 const io = require('socket.io')(http);
 const path = require('path');
 const bodyParser = require('body-parser');
-const testGenerator = require('./test_generator');
+const TestContainer = require('./TestContainer');
 const opn = require('opn');
 
 const PORT = 4050;
@@ -25,7 +25,7 @@ const generateRequests = (testData) => {
     concurrency: 100,
   };
 
-  testGenerator.test(config, (error, result) => {
+  TestContainer.test(config, (error, result) => {
     if (error) {
       return callback(error);
     }
